feat: support a client system init
